even odds

+++ codeforces/evenodds.cpp
@@ -2,17 +2,17 @@
 using namespace std;
 
 int main() {
-    long long n, k;
+    int n, k;
     cin >> n >> k;
 
-    long long odd_count;
+    int odd_count;
     if (n % 2 == 1) {
         odd_count = (n + 1) / 2;
     } else {
         odd_count = n / 2;
     }
 
-    long long result;
+    int result;
     if (k <= odd_count) {
         result = 2 * k - 1;
     } else {
